+++ src/middlewares/upload-to-gist-or-highlight.ts
@@ -1,7 +1,7 @@
 import { Composer } from '../composer';
 import { OnMiddleware } from '../types';
 import { botHasSufficientPermissions, isGroupChat } from '../guards';
-import { getCodeFromMessage, runEnry, updloadToGist } from '../utils';
+import { getCodeFromMessage, runEnry, uploadToGist } from '../utils';
 import { log } from '../logger';
 import { link, userMention } from '../utils/html';
 import { GIST_UPLOAD_LINE_COUNT_THRESHOLD } from '../constants';
@@ -15,9 +15,10 @@ export const uploadToGistOrHighlight = Composer.guardAll(
     if (!sourceCode) return next();
     const lineCount = sourceCode.split('\n').length;
     if (lineCount < GIST_UPLOAD_LINE_COUNT_THRESHOLD) return next();
-    const languageName = await runEnry(sourceCode);
-    log.info('Uploading to gist a chunk of `%s` code', languageName);
-    const codeUrl = await updloadToGist(languageName, sourceCode);
+    const enryResult = await runEnry(sourceCode);
+    if (!enryResult) return next();
+    log.info('Uploading to gist a chunk of `%s` code', enryResult.language);
+    const codeUrl = await uploadToGist(enryResult, sourceCode);
     if (!codeUrl) return next();
     await ctx.deleteMessage();
     const linkWithText = link(codeUrl, 'GitHub Gist');
